Password Reset Implementation
First implementation of the password reset functionality, as an API,
using POST requests. The link sent via email to request a password reset
implies that a page parses the url via JS and submits a POST request.
This is because we need to ask the user for their new password anyway.

In the meantime, add databaseText() to the error log when a query fails
instead of just having driverText(), and do not save sshkey to the DB to
avoid failures for a missing optional parameter: we are not using sshkey
anyway!

References #1